Add initial support for delete summary page.
