+++ workbaskets/jinja2/workbaskets/delete_changes.jinja
@@ -1,0 +1,3 @@
+{% extends "layouts/layout.jinja" %}
+
+testing
